+++ labstep/entities/collaborator/model.py
@@ -11,7 +11,7 @@ class Collaborator(Entity):
 
     def unassign (self):
         """
-        Unassign collabortor.
+        Unassign collaborator.
 
         Example
         -------
@@ -21,4 +21,3 @@ class Collaborator(Entity):
             collaborators[0].unassign()
         """
         return collaboratorRepository.unassign(self)
-