+++ labstep/service/htmlToPDF.py
@@ -3,10 +3,11 @@
 # Author: Labstep <dev@labstep.com>
 
 from labstep.service.request import requestService
-import base64
+from labstep.constants import UNSPECIFIED
+import json
 
 
-def htmlToPDF(authenticatedUser, html):
+def htmlToPDF(authenticatedUser, html=UNSPECIFIED,file_id=UNSPECIFIED,returnFile=False):
     """
     Converts HTML into the JSON format used in protocols
     and experiment entries.
@@ -21,9 +22,28 @@ def htmlToPDF(authenticatedUser, html):
     """
     headers = {"Authorization": f"Bearer {authenticatedUser.token}"}
 
-    body = {"html": html}
+    if html is UNSPECIFIED and file_id is UNSPECIFIED:
+        raise ValueError("html or file_id must be provided")
+
+    body = {"group_id": authenticatedUser.activeWorkspace}
+
+    if html is not UNSPECIFIED:
+        body["html"] = html
+
+    if file_id is not UNSPECIFIED:
+        body["file_id"] = file_id
 
     url = "https://pdf-generator.labstep.com"
     response = requestService.post(url, json=body, headers=headers)
 
-    return base64.b64decode(response.content)
+    fileID = json.loads(response.content)["file_id"]
+    file = authenticatedUser.getFile(fileID)
+
+    file.save()
+
+
+    if returnFile:
+        return file
+    else:
+        data = file.getData()
+        return data