+++ tests/functional/shared.py
@@ -1,17 +1,18 @@
-from .fixtures import fixtures, testString, tableData
-from labstep.service.request import RequestException
-from labstep.generic.entity.repository import newEntity
-from labstep.entities.protocol.model import Protocol
-from labstep.entities.experiment.model import Experiment
-from labstep.entities.tag.model import Tag
+from labstep.entities.collection.model import Collection
 from labstep.entities.device.model import Device
+from labstep.entities.experiment.model import Experiment
+from labstep.entities.orderRequest.model import OrderRequest
+from labstep.entities.protocol.model import Protocol
 from labstep.entities.resource.model import Resource
-from labstep.entities.workspace.model import Workspace
 from labstep.entities.resourceCategory.model import ResourceCategory
 from labstep.entities.resourceItem.model import ResourceItem
 from labstep.entities.resourceLocation.model import ResourceLocation
-from labstep.entities.orderRequest.model import OrderRequest
-from labstep.entities.collection.model import Collection
+from labstep.entities.tag.model import Tag
+from labstep.entities.workspace.model import Workspace
+from labstep.generic.entity.repository import newEntity
+from labstep.service.request import RequestException
+
+from .fixtures import fixtures, tableData, testString
 
 
 class SharedTests:
@@ -114,7 +115,7 @@ class SharedTests:
         return jupyter_notebook_entity.id == jupyter_notebook.id \
             and jupyter_notebook_entity.name == "test"
 
-    def conditions(self, entity):
+    def protocolConditions(self, entity):
         user = fixtures.defaultUser()
         test_resource = user.newResource(name='Test')
         output_data_field = entity.addDataField(
@@ -127,6 +128,27 @@ class SharedTests:
         condition_entity = condition[0]
 
         condition_from_get = entity.getConditions()[0]
+
+        input_data_field_from_get = condition_entity.getDataFields()
+        input_inventory_field_from_get = condition_entity.getInventoryFields()
+
+        return condition_entity.guid == condition_from_get.guid \
+            and condition_entity.guid == input_data_field_from_get[0].protocol_condition_guid \
+            and condition_entity.guid == input_inventory_field_from_get[0].protocol_condition_guid
+
+    def experimentConditions(self, entity):
+        user = fixtures.defaultUser()
+        test_resource = user.newResource(name='Test')
+        output_data_field = entity.addDataField(
+            'Test', extraParams={'is_variable': True, 'is_output': True})
+        output_inventory_field = entity.addInventoryField('Test',
+                                                          resource_id=test_resource.id,
+                                                          extraParams={'is_variable': True, 'is_output': True})
+
+        condition = entity.addConditions(1)
+        condition_entity = condition[0]
+
+        condition_from_get = entity.getConditions()[1]
 
         input_data_field_from_get = condition_entity.getDataFields()
         input_inventory_field_from_get = condition_entity.getInventoryFields()